+++ backend/crm/src/main/java/cn/cordys/crm/opportunity/controller/OpportunityFollowPlanController.java
@@ -1,8 +1,12 @@
 package cn.cordys.crm.opportunity.controller;
 
 import cn.cordys.common.constants.FormKey;
+import cn.cordys.common.constants.FormKeyConstants;
+import cn.cordys.common.constants.ModuleKey;
 import cn.cordys.common.constants.PermissionConstants;
 import cn.cordys.common.pager.PagerWithOption;
+import cn.cordys.common.permission.CsPermission;
+import cn.cordys.common.util.BeanUtils;
 import cn.cordys.common.utils.ConditionFilterUtils;
 import cn.cordys.context.OrganizationContext;
 import cn.cordys.crm.follow.domain.FollowUpPlan;
@@ -17,7 +21,6 @@ import cn.cordys.security.SessionUtils;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.annotation.Resource;
-import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
@@ -32,57 +35,61 @@ public class OpportunityFollowPlanController {
     private FollowUpPlanService followUpPlanService;
 
     @PostMapping("/add")
-    @RequiresPermissions(PermissionConstants.OPPORTUNITY_MANAGEMENT_UPDATE)
     @Operation(summary = "添加项目跟进计划")
     public FollowUpPlan add(@Validated @RequestBody FollowUpPlanAddRequest request) {
+        followUpPlanService.checkPlanPermission(BeanUtils.copyBean(new FollowUpPlan(), request),
+                OrganizationContext.getOrganizationId(), SessionUtils.getUserId(), false);
         return followUpPlanService.add(request, SessionUtils.getUserId(), OrganizationContext.getOrganizationId());
     }
 
 
     @PostMapping("/update")
-    @RequiresPermissions(PermissionConstants.OPPORTUNITY_MANAGEMENT_UPDATE)
     @Operation(summary = "更新项目跟进计划")
     public FollowUpPlan update(@Validated @RequestBody FollowUpPlanUpdateRequest request) {
+        followUpPlanService.checkUpdatePermission(request.getId(), SessionUtils.getUserId());
         return followUpPlanService.update(request, SessionUtils.getUserId(), OrganizationContext.getOrganizationId());
     }
 
 
     @PostMapping("/page")
-    @RequiresPermissions(PermissionConstants.OPPORTUNITY_MANAGEMENT_READ)
     @Operation(summary = "项目跟进计划列表")
     public PagerWithOption<List<FollowUpPlanListResponse>> list(@Validated @RequestBody FollowUpPlanPageRequest request) {
+        FollowUpPlan followPlanRecord = new FollowUpPlan();
+        followPlanRecord.setOpportunityId(request.getSourceId());
+        followPlanRecord.setType(ModuleKey.CUSTOMER.name());
+        followUpPlanService.checkPlanPermission(followPlanRecord, OrganizationContext.getOrganizationId(), SessionUtils.getUserId(), true);
         ConditionFilterUtils.parseCondition(request, FormKey.FOLLOW_PLAN.getKey());
-        return followUpPlanService.list(request, SessionUtils.getUserId(), OrganizationContext.getOrganizationId(), "OPPORTUNITY", "CUSTOMER", null);
+        return followUpPlanService.list(request, SessionUtils.getUserId(), OrganizationContext.getOrganizationId(), "OPPORTUNITY", "CUSTOMER");
     }
 
 
     @GetMapping("/get/{id}")
-    @RequiresPermissions(PermissionConstants.OPPORTUNITY_MANAGEMENT_READ)
     @Operation(summary = "项目跟进计划详情")
     public FollowUpPlanDetailResponse get(@PathVariable String id) {
+        followUpPlanService.checkPlanPermission(id, OrganizationContext.getOrganizationId(), SessionUtils.getUserId(), true);
         return followUpPlanService.get(id, OrganizationContext.getOrganizationId());
     }
 
 
     @GetMapping("/cancel/{id}")
-    @RequiresPermissions(PermissionConstants.OPPORTUNITY_MANAGEMENT_UPDATE)
     @Operation(summary = "取消项目跟进计划")
     public void cancelPlan(@PathVariable String id) {
+        followUpPlanService.checkUpdatePermission(id, SessionUtils.getUserId());
         followUpPlanService.cancelPlan(id, SessionUtils.getUserId());
     }
 
 
     @GetMapping("/delete/{id}")
     @Operation(summary = "项目删除跟进计划")
-    @RequiresPermissions(PermissionConstants.OPPORTUNITY_MANAGEMENT_UPDATE)
     public void deletePlan(@PathVariable String id) {
+        followUpPlanService.checkUpdatePermission(id, SessionUtils.getUserId());
         followUpPlanService.delete(id);
     }
 
     @PostMapping("/status/update")
-    @RequiresPermissions(PermissionConstants.OPPORTUNITY_MANAGEMENT_UPDATE)
     @Operation(summary = "项目更新跟进计划状态")
     public void updateStatus(@Validated @RequestBody FollowUpPlanStatusRequest request) {
+        followUpPlanService.checkUpdatePermission(request.getId(), SessionUtils.getUserId());
         followUpPlanService.updateStatus(request, SessionUtils.getUserId());
     }
